Move AppBarStatus to its own file
