+++ WooCommerce/src/main/kotlin/com/woocommerce/android/ui/base/BaseFragment.kt
@@ -2,11 +2,10 @@ package com.woocommerce.android.ui.base
 
 import android.os.Bundle
 import android.view.View
-import androidx.annotation.DrawableRes
 import androidx.annotation.LayoutRes
 import androidx.fragment.app.Fragment
-import com.woocommerce.android.R
 import com.woocommerce.android.ui.dialog.WooDialog
+import com.woocommerce.android.ui.main.AppBarStatus
 import com.woocommerce.android.ui.main.MainActivity
 import com.woocommerce.android.viewmodel.MultiLiveEvent.Event.ShowDialog
 
@@ -82,14 +81,4 @@ open class BaseFragment : Fragment, BaseFragmentView {
             negBtnAction = this.negativeBtnAction
         )
     }
-}
-
-sealed class AppBarStatus {
-    object Hidden: AppBarStatus()
-    data class Visible(
-        @DrawableRes
-        val navigationIcon: Int? = R.drawable.ic_back_24dp,
-        val hasShadow: Boolean = true,
-        val hasDivider: Boolean = false
-    ): AppBarStatus()
 }
