+++ WooCommerce/src/main/kotlin/com/woocommerce/android/ui/main/AppBarStatus.kt
@@ -1,0 +1,14 @@
+package com.woocommerce.android.ui.main
+
+import androidx.annotation.DrawableRes
+import com.woocommerce.android.R
+
+sealed class AppBarStatus {
+    object Hidden: AppBarStatus()
+    data class Visible(
+        @DrawableRes
+        val navigationIcon: Int? = R.drawable.ic_back_24dp,
+        val hasShadow: Boolean = true,
+        val hasDivider: Boolean = false
+    ): AppBarStatus()
+}
