+++ WooCommerce/src/main/kotlin/com/woocommerce/android/ui/orders/details/PrintingInstructionsFragment.kt
@@ -1,8 +1,8 @@
 package com.woocommerce.android.ui.orders.details
 
 import com.woocommerce.android.R
-import com.woocommerce.android.ui.base.AppBarStatus
 import com.woocommerce.android.ui.base.BaseFragment
+import com.woocommerce.android.ui.main.AppBarStatus
 import dagger.hilt.android.AndroidEntryPoint
 
 @AndroidEntryPoint
