+++ WooCommerce/src/main/kotlin/com/woocommerce/android/ui/orders/shippinglabels/LabelFormatOptionsFragment.kt
@@ -2,8 +2,8 @@ package com.woocommerce.android.ui.orders.shippinglabels
 
 import com.woocommerce.android.R
 import com.woocommerce.android.analytics.AnalyticsTracker
-import com.woocommerce.android.ui.base.AppBarStatus
 import com.woocommerce.android.ui.base.BaseFragment
+import com.woocommerce.android.ui.main.AppBarStatus
 
 class LabelFormatOptionsFragment : BaseFragment(R.layout.fragment_label_format_options) {
     override val activityAppBarStatus: AppBarStatus
